FinApp/requirements.txt
alabaster==0.7.13
asgiref==3.7.2
Babel==2.13.1
certifi==2023.7.22
charset-normalizer==3.3.2
colorama==0.4.6
Django==4.2.6
docutils==0.20.1
idna==3.4
imagesize==1.4.1
Jinja2==3.1.2
MarkupSafe==2.1.3
nulltype==2.3.1
packaging==23.2
plaid-python==17.0.0
Pygments==2.16.1
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
snowballstemmer==2.2.0
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
